feat: Refactored sdk to remove event object entirely
